+++ crates/holochain/src/core/sys_validate/error.rs
@@ -210,6 +210,8 @@ pub enum PrevActionErrorKind {
     InvalidRoot,
     #[error("Root of source chain must have a timestamp greater than the Dna's origin_time")]
     InvalidRootOriginTime,
+    #[error("No more actions are allowed after a chain close")]
+    ActionAfterChainClose,
     #[error("Previous action sequence number {1} != ({0} - 1)")]
     InvalidSeq(u32, u32),
     #[error("Action is not the first, so needs previous action")]